requirements.txt
Package                 Version     Editable project location
----------------------- ----------- ---------------------------------------------
absl-py                 2.3.1
ansitable               0.11.4
astunparse              1.6.3
certifi                 2025.11.12
cfgv                    3.4.0
charset-normalizer      3.4.4
chess                   1.11.2
colored                 2.3.1
contourpy               1.3.3
csc376_franky           0.0.0
cycler                  0.12.1
distlib                 0.4.0
filelock                3.20.0
flatbuffers             25.9.23
fonttools               4.60.1
gast                    0.6.0
google-pasta            0.2.0
grpcio                  1.76.0
h5py                    3.15.1
identify                2.6.15
idna                    3.11
keras                   3.12.0
kiwisolver              1.4.9
libclang                18.1.1
Markdown                3.10
markdown-it-py          4.0.0
MarkupSafe              3.0.3
matplotlib              3.8.4
mdurl                   0.1.2
ml_dtypes               0.5.4
msl-loadlib             0.10.0
namex                   0.1.0
nodeenv                 1.9.1
numpy                   1.26.4
opencv-python           4.11.0.86
opt_einsum              3.4.0
optree                  0.18.0
packaging               25.0
pandas                  2.3.3
pgraph-python           0.6.3
pillow                  12.0.0
pip                     25.2
platformdirs            4.5.0
pre_commit              4.3.0
progress                1.6.1
protobuf                6.33.1
pybind11                3.0.1
pydraughts              0.6.7
Pygments                2.19.2
pyparsing               3.2.5
python-dateutil         2.9.0.post0
pytz                    2025.2
PyYAML                  6.0.3
requests                2.32.5
rich                    14.2.0
roboticstoolbox-python  1.1.1
rtb-data                1.0.1
ruckig                  0.15.3
scipy                   1.16.3
setuptools              80.9.0
six                     1.17.0
spatialgeometry         1.1.0
spatialmath-python      1.1.14
swift-sim               1.1.0       /home/teachinglab_student/csc376/swift-csc376
tensorboard             2.20.0
tensorboard-data-server 0.7.2
tensorflow              2.20.0
termcolor               3.2.0
typing_extensions       4.15.0
tzdata                  2025.2
urllib3                 2.5.0
virtualenv              20.35.4
websockets              13.0.1
Werkzeug                3.1.3
wheel                   0.45.1
wrapt                   2.0.1
